+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added 'get_summary_variables' to return dictionary of computed summary variables ([#4824](https://github.com/pybamm-team/PyBaMM/pull/4824))
 - Added support for particle size distributions combined with particle mechanics. ([#4807](https://github.com/pybamm-team/PyBaMM/pull/4807))
 - Added InputParameter support in PyBamm experiments ([#4826](https://github.com/pybamm-team/PyBaMM/pull/4826))
+- Added pchip Interpolator ([#4871](https://github.com/pybamm-team/PyBaMM/pull/4871))
 
 ## Breaking changes